requirements.txt
anyio==4.0.0
async-timeout==4.0.3
asyncclick==8.1.3.4
certifi==2023.7.22
charset-normalizer==3.2.0
exceptiongroup==1.1.3
idna==3.4
pkcs7==0.1.2
pycryptodome==3.19.0
pydantic==1.10.12
PyP100==0.1.2
python-dotenv==1.0.0
python-kasa==0.5.3
pytz==2023.3.post1
requests==2.31.0
schedule==1.2.0
sniffio==1.3.0
typing-extensions==4.8.0
urllib3==2.0.5
